refactor: improve password strength component with dynamic props and logic
- Moved password strength calculation and requirement logic into the PasswordStrength component
- Simplified props by passing raw password instead of pre-calculated values
- Removed redundant methods from parent component
- Maintained existing functionality with more encapsulated component design
